refactor: refactor by adding new controller
