+++ .env.local
@@ -1,5 +1,5 @@
 APP_NAME="[Local] Sistema Informativo Nomadelfia"
-APP_KEY=
+APP_KEY=base64:wvPNoQMla80AG/g+oSoXm0/OGKoAXpV4sG6igELioTg=
 APP_URL=http://127.0.0.1:8080
 APP_ENV=local
