+++ resources/views/nomadelfia/index.blade.php
@@ -30,7 +30,7 @@
         </a>
         <div class="dropdown-menu" aria-labelledby="navbarPesone">
             <a class="dropdown-item" href="{{ route('nomadelfia.persone.ricerca') }}">Ricerca persone</a>
-            <a class="dropdown-item" href="{{ route('nomadelfia.persone.inserimento')}}">Inserisci Persona</a>
+            <a class="dropdown-item" href="{{ route('nomadelfia.persone.create')}}">Inserisci Persona</a>
         </div>
     </li>
