+++ src/App/Sin/Nomadelfia/PopolazioneNomadelfia/Controllers/PopolazioneNomadelfiaController.php
@@ -15,30 +15,6 @@ use Spatie\Activitylog\Models\Activity;
 
 class PopolazioneNomadelfiaController extends CoreBaseController
 {
-    public function index()
-    {
-        $totale = PopolazioneNomadelfia::totalePopolazione();
-        $maggiorenni = PopolazioneNomadelfia::maggiorenni();
-        $effettivi = PopolazioneNomadelfia::effettivi();
-        $postulanti = PopolazioneNomadelfia::postulanti();
-        $ospiti = PopolazioneNomadelfia::ospiti();
-        $sacerdoti = PopolazioneNomadelfia::sacerdoti();
-        $mvocazione = PopolazioneNomadelfia::mammeVocazione();
-        $nomanamma = PopolazioneNomadelfia::nomadelfaMamma();
-        $figliMaggiorenni = PopolazioneNomadelfia::figliMaggiorenni();
-        $minorenni = PopolazioneNomadelfia::figliMinorenni();
-        $stats = PopolazioneNomadelfia::stats();
-
-        $figli = PopolazioneNomadelfia::byPosizione('FIGL');
-
-        $gruppi = GruppoFamiliare::countComponenti();
-        $posizioniFamiglia = PopolazioneNomadelfia::posizioneFamigliaCount();
-        $famiglieNumerose = Famiglia::famiglieNumerose();
-
-        $activities = Activity::inLog('nomadelfia')->orderBy('created_at', 'DESC')->take(20)->get();
-
-        return view('nomadelfia.summary', compact('totale', 'maggiorenni', 'effettivi', 'postulanti', 'ospiti', 'sacerdoti', 'mvocazione', 'nomanamma', 'figliMaggiorenni', 'minorenni', 'figli', 'gruppi', 'posizioniFamiglia', 'famiglieNumerose', 'stats', 'activities'));
-    }
 
     public function show()
     {
